Revert "fix(staff): fix staff getAll query params types and checks; add tests…"
