+++ test/staff/get-all.test.ts
@@ -2,8 +2,21 @@ import * as dotenv from 'dotenv'
 import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
 dotenv.config()
-import { v0 } from '../../src/tillhub-js'
-import { initThInstance } from '../util'
+import { TillhubClient, v0 } from '../../src/tillhub-js'
+
+let user = {
+  username: 'test@example.com',
+  password: '12345678',
+  clientAccount: 'someuuid',
+  apiKey: '12345678'
+}
+
+if (process.env.SYSTEM_TEST) {
+  user.username = process.env.SYSTEM_TEST_USERNAME || user.username
+  user.password = process.env.SYSTEM_TEST_PASSWORD || user.password
+  user.clientAccount = process.env.SYSTEM_TEST_CLIENT_ACCOUNT_ID || user.clientAccount
+  user.apiKey = process.env.SYSTEM_TEST_API_KEY || user.apiKey
+}
 
 const legacyId = '4564'
 
@@ -39,7 +52,21 @@ describe('v0: Staff: can get all', () => {
       })
     }
 
-    const th = await initThInstance()
+    const options = {
+      credentials: {
+        username: user.username,
+        password: user.password
+      },
+      base: process.env.TILLHUB_BASE
+    }
+
+    const th = new TillhubClient()
+
+    th.init(options)
+    await th.auth.loginUsername({
+      username: user.username,
+      password: user.password
+    })
 
     const Staff = th.staff()
 
@@ -72,85 +99,26 @@ describe('v0: Staff: can get all', () => {
         })
     }
 
+    const options = {
+      credentials: {
+        username: user.username,
+        password: user.password
+      },
+      base: process.env.TILLHUB_BASE
+    }
+
+    const th = new TillhubClient()
+
+    th.init(options)
+    await th.auth.loginUsername({
+      username: user.username,
+      password: user.password
+    })
+
     try {
-      const th = await initThInstance()
       await th.staff().getAll()
     } catch (err) {
       expect(err.name).toBe('StaffFetchFailed')
     }
-  })
-
-  it('accepts query strings', async () => {
-    const start = '123'
-    const deleted = false
-    const active = true
-    const qs = { start, deleted, active }
-    if (process.env.SYSTEM_TEST !== 'true') {
-      mock.onPost('https://api.tillhub.com/api/v0/users/login').reply(function (config) {
-        return [
-          200,
-          {
-            token: '',
-            user: {
-              id: '123',
-              legacy_id: legacyId
-            }
-          }
-        ]
-      })
-
-      mock
-        .onGet(
-          `https://api.tillhub.com/api/v0/staff/${legacyId}?start=${start}&deleted=${deleted}&active=${active}`
-        )
-        .reply(function (config) {
-          return [
-            200,
-            {
-              count: 1,
-              results: [{}]
-            }
-          ]
-        })
-    }
-
-    const th = await initThInstance()
-    const { data } = await th.staff().getAll(qs)
-    expect(Array.isArray(data)).toBe(true)
-  })
-
-  it('ignore invalid query strings', async () => {
-    const start = ''
-    const deleted = undefined
-    const active = undefined
-    const qs = { start, deleted, active }
-    if (process.env.SYSTEM_TEST !== 'true') {
-      mock.onPost('https://api.tillhub.com/api/v0/users/login').reply(function (config) {
-        return [
-          200,
-          {
-            token: '',
-            user: {
-              id: '123',
-              legacy_id: legacyId
-            }
-          }
-        ]
-      })
-
-      mock.onGet(`https://api.tillhub.com/api/v0/staff/${legacyId}`).reply(function (config) {
-        return [
-          200,
-          {
-            count: 1,
-            results: [{}]
-          }
-        ]
-      })
-    }
-
-    const th = await initThInstance()
-    const { data } = await th.staff().getAll(qs)
-    expect(Array.isArray(data)).toBe(true)
   })
 })
